Lazy loading and collapsed messages
This commit adds asynchronous guild icon loading as well as lazy loading
of channel messages. This means that channel messages are not loaded
until the user actually goes to that channel.

The commit also adds ningen support into Fyne, which includes a plain
text Markdown renderer as well as a lot more abstractions that will be
useful in the future.

Token login is also added into the login dialog as a button.